+++ packages/components/src/Form/Inputs/InlineTextArea/index.ts
@@ -24,17 +24,4 @@
 
  */
 
-export * from './Checkbox'
-export * from './Combobox'
 export * from './InlineTextArea'
-export * from './InputChips'
-export * from './InputDate'
-export * from './InputDateRange'
-export * from './InputHidden'
-export * from './InputSearch'
-export * from './InputText'
-export * from './Radio'
-export * from './Select'
-export * from './Slider'
-export * from './TextArea'
-export * from './ToggleSwitch'