+++ packages/components/src/Form/Fields/FieldTextArea/FieldTextArea.tsx
@@ -24,45 +24,34 @@
 
  */
 
-export * from './ActionList'
-export * from './Avatar'
-export * from './Badge'
-export * from './Banner'
-export * from './Button'
-export * from './Calendar'
-export * from './Card'
-export * from './Chip'
-export * from './DateTimeFormat'
-export * from './Divider'
-export * from './Form'
-export * from './GlobalStyle'
-export * from './Icon'
-export * from './Layout'
-export * from './Link'
-export * from './List'
-export * from './Menu'
-export * from './Modal'
-export * from './Popover'
-export * from './Spinner'
-export * from './Sidebar'
-export * from './Table'
-export * from './Tabs'
-export * from './Tooltip'
-export * from './Text'
-export * from './VisuallyHidden'
-export * from './utils'
+import React, { FC } from 'react'
+import styled from 'styled-components'
+import { v4 as uuid } from 'uuid'
+import { useFormContext } from '../../Form'
+import { TextArea, TextAreaProps } from '../../Inputs/TextArea'
+import { Field, FieldProps, omitFieldProps, pickFieldProps } from '../Field'
 
-/** Provided here for backwards compatibility.
- * @TODO - Remove before 1.0
- **/
+export interface FieldTextAreaProps extends FieldProps, TextAreaProps {}
 
-export {
-  palette,
-  radii,
-  semanticColors,
-  SemanticColor,
-  SemanticColors,
-  SpacingSizes,
-  theme,
-  Theme,
-} from '@looker/design-tokens'
+const FieldTextAreaComponent: FC<FieldTextAreaProps> = ({ ...props }) => {
+  const { id = uuid() } = props
+  const validationMessage = useFormContext(props)
+  return (
+    <Field
+      id={id}
+      alignValidationMessage="bottom"
+      validationMessage={validationMessage}
+      {...pickFieldProps(props)}
+    >
+      <TextArea
+        {...omitFieldProps(props)}
+        validationType={validationMessage && validationMessage.type}
+      />
+    </Field>
+  )
+}
+
+FieldTextAreaComponent.displayName = 'FieldTextAreaComponent'
+
+export const FieldTextArea = styled(FieldTextAreaComponent)``
+FieldTextArea.defaultProps = { width: '13rem' }