+++ CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `TextArea component was created with documentation and tests
-- `FieldTextArea component was created with documentation and tests
+- `TextArea` component was created with documentation and tests
+- `FieldTextArea` component was created with documentation and tests
 
 ## [0.7.24] - 2020-03-12