+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `TextArea` component was created with documentation and tests
 - `FieldTextArea` component was created with documentation and tests
+- `InlineTextArea` component was created with documentation and tests
 
 ## [0.7.24] - 2020-03-12