+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Change
 
-move path for InlineInputText now under Form/Inputs
+- move path for InlineInputText now under Form/Inputs
+- update InputSearch to display search icon
 
 ## [0.7.24] - 2020-03-12